fix set_Ms

+++ src/fem/mesh.jl
@@ -56,16 +56,16 @@ function load_mesh_netgen_neutral(fname::String)
 
     N = parse(Int64, readline(io)) #then it sould be the total number of cells
     mesh.number_cells = N
-    mesh.cell_verts = zeros(Float64, 4, N)
-    mesh.region_ids = ones(Integer, N)
+    mesh.cell_verts = zeros(Int32, 4, N)
+    mesh.region_ids = ones(Int32, N)
 
     for i in 1:N
         x = split(readline(io))
-        mesh.region_ids[i] = parse(Int64, x[1])
-        mesh.cell_verts[1, i] = parse(Int64, x[2])
-        mesh.cell_verts[2, i] = parse(Int64, x[3])
-        mesh.cell_verts[3, i] = parse(Int64, x[4])
-        mesh.cell_verts[4, i] = parse(Int64, x[5])
+        mesh.region_ids[i] = parse(Int32, x[1])
+        mesh.cell_verts[1, i] = parse(Int32, x[2])
+        mesh.cell_verts[2, i] = parse(Int32, x[3])
+        mesh.cell_verts[3, i] = parse(Int32, x[4])
+        mesh.cell_verts[4, i] = parse(Int32, x[5])
     end
 
     N = parse(Int64, readline(io)) #finally it sould be the total surface number
@@ -123,7 +123,7 @@ function build_boundary_surfaces!(mesh::FEMesh)
     end
 
     mesh.number_faces_bnd = length(face_info)
-    mesh.face_verts = zeros(Int64, 3, mesh.number_faces_bnd)
+    mesh.face_verts = zeros(Int32, 3, mesh.number_faces_bnd)
 
     # Process each boundary face to ensure correct normal orientation
     for (i, (key, (_, opp))) in enumerate(face_info)
